# Optimize find_position performance