@Horusiath 2022-07-22
Thanks @ArsnealX - we didn't work on optimising find_position atm. simply because for client apps, insert operations are bound to user reaction time: it's highly doubtful that any human being is capable to making 12000 key strokes in 70 secs.

Eventually, we'll planning on implementing 2 ways to improve the speed of insert ops:

1. Implement and expose cursors API (it's more advanced idea of search markers, conceptually similar to how cursors in text editors work), which are structure that refers to an actual position in the text/array and can advance together with user insert/delete operations.
2. Reimplement `ClientBlockList` as an rope data structure, which has logarithmic random lookup complexity (current impl. has linear).