@ArsnealX 2022-07-22
@Horusiath Thank you for keeping me in the loop. I'm looking forward to these changes coming soon. 

The benchmarks are coming from our user case in the production app:

 - When we init an existing rich-text document to ydoc, we need to split the text with different attributes to multiply insert operations for future editing.

- The user can simply copy and paste a huge text with complex attributes into the document. 

- Some of our users have a large to-do (approx 10000+ lines) in a document, the larger the doc, the worse the performance.

Maybe there are alternative ways to bypass the massive insert operations to create a new ydoc, we're looking forward to your suggestions.